backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
psycopg2-binary==2.9.9

# Cache / Queue
redis==5.0.8
celery==5.4.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
python-dotenv==1.0.1

# HTTP
httpx==0.27.2
aiohttp==3.10.5

# Platform SDKs
facebook-business==20.0.2
google-ads==24.1.0
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
salesforce-marketing-cloud-sdk==1.3.0

# Claude AI
anthropic==0.34.2

# Data processing
pandas==2.2.3
pydantic==2.9.2
pydantic-settings==2.5.2

# Logging & monitoring
structlog==24.4.0
